requirements.txt
crewai>=0.11.2
crewai-tools>=0.1.6
langchain-community>=0.0.29
python-dotenv==1.0.0
rich==13.7.0
openai>=1.13.3 
utils